+++ CHANGELOG.md
@@ -12,7 +12,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
  * NOTE: `/api/packages/<package>` calls only use `gzip` cache, which may increase
          CPU utilization if there are many clients that do no accept the `gzip` content-encoding.
  * NOTE: Started to use cached package list for all use case where a list of package versions is required.
- * NOTE: started to populate and use `Package.versionsCount`.
+ * NOTE: started to populate and use `Package.versionCount`.
          TODO(deferred): add integrity check in next release.
 
 ## `20211013t105500-all`